+++ equi7grid_lite/main.py
@@ -1,17 +1,19 @@
 import math
+import multiprocessing as mp
 import re
-from typing import Dict, List, Literal, Optional, Tuple, Union
+from functools import partial
+from typing import Dict, List, Literal, Optional, Tuple
 
 import geopandas as gpd
 import numpy as np
 import pandas as pd
+import pyproj
+import shapely.geometry
 import shapely.wkt
 from nptyping import NDArray, Shape
-from shapely.geometry import Point
 
-from equi7grid_lite.dataclass import (Equi7GridZone, InternalDataset,
-                                      dict_to_datamodel)
-from equi7grid_lite.utils import load_grid, utils_add_id, haversine_distance
+from equi7grid_lite.dataclass import Equi7GridZone, InternalDataset, dict_to_datamodel
+from equi7grid_lite.utils import get_id, haversine_distance, intersects_func, load_grid
 
 
 class Equi7Grid:
@@ -34,8 +36,8 @@ class Equi7Grid:
         Initialize the `Equi7Grid` class.
 
         Args:
-            min_grid_size (int, optional): The size of the grid tile at the finest level,
-                measured in meters. Defaults to 2560.
+            min_grid_size (int, optional): The size of the grid tile at the finest
+                level, measured in meters. Defaults to 2560.
 
         Attributes:
             min_grid_size (int): The size of the grid tile at the finest level.
@@ -50,6 +52,8 @@ class Equi7Grid:
             zone_geometry_max_grids (Dict[str, gpd.GeoDataFrame]): The maximum grid size per
                 zone, used as a reference to create subgrids. Grids larger than this size
                 cannot be created.
+            zone_origin (List[Tuple[float, float]]): The origin of each
+                Equi7Grid zone.
             max_grid_size (int): The maximum grid size in meters.
         """
 
@@ -70,7 +74,6 @@ class Equi7Grid:
         # Calculate the levels
         max_level: int = int(np.log2(self.max_grid_size / self.min_grid_size))
         self.levels = list(range(max_level))
-
 
         # Define the Equi7Grid zones metadata
         self.AN: Equi7GridZone = dict_to_datamodel(zone_metadata, 0)
@@ -146,8 +149,14 @@ class Equi7Grid:
 
             # Create the GeoDataFrame
             grid = gpd.GeoDataFrame(geometry=tiles, crs=crs)
-            grid["zone"] = name
-            grid["level"] = max_level_order
+
+            # Add the zone and level
+            grid.insert(loc=0, column="zone", value=name)
+            grid.insert(loc=1, column="level", value=max_level_order)
+
+            # Add the land column
+            land = self.zone_land[self.zone_names.index(name)]
+            grid.insert(loc=2, column="land", value=grid.intersects(land).values)
             zone_containers[name] = grid
 
         return zone_containers, max_grid_size
@@ -156,10 +165,9 @@ class Equi7Grid:
         self,
         level: int,
         zone: Literal["AN", "NA", "OC", "SA", "AF", "EU", "AS"],
-        mask: Optional[shapely.geometry.base.BaseGeometry] = None,
-        coverland: Optional[bool] = True
+        mask: Optional[gpd.GeoSeries] = None,
     ) -> gpd.GeoDataFrame:
-        """Create a grid for a specific zone.
+        """Create a Equi7Grid polygon for a specific zone.
 
         Args:
             level (int): The level of the grid. The level must be
@@ -167,9 +175,7 @@ class Equi7Grid:
                 the `create_init_grid` method.
             zone (Literal["AN", "NA", "OC", "SA", "AF", "EU", "AS"]): The
                 Equi7Grid zone.
-            coverland (bool, optional): If True, the grid system will be created
-                only for the land area. Defaults to True.
-            mask (Optional[shapely.geometry.base.BaseGeometry], optional): If
+            mask (Optional[gpd.GeoSeries], optional): The mask to apply
                 provided, the grid will be created only within the mask.
                 Defaults to None.
 
@@ -187,21 +193,22 @@ class Equi7Grid:
             )
 
         # Obtain the index of the zone in the metadata
-        zone_index = self.zone_names.index(zone)
+        zone_index: int = self.zone_names.index(zone)
 
         # Define the zone geometry
-        zone_geometry = self.zone_geometry[zone_index]
-
-        # Intersect the zone with the land if land is True
-        if coverland:
-            land_geometry = self.zone_land[zone_index]
-            zone_geometry = zone_geometry.intersection(land_geometry)
+        zone_geometry: shapely.geometry.Polygon = self.zone_geometry[zone_index]
 
         # Intersect the zone with the mask if mask is not None
         if mask is not None:
-            mask_gpd = gpd.GeoDataFrame(geometry=[mask], crs="EPSG:4326")
-            mask_plane = mask_gpd.to_crs(self.zone_crs[zone_index]).geometry[0]
-            zone_geometry = zone_geometry.intersection(mask_plane)
+            # Always reset the index
+            mask.reset_index(drop=True, inplace=True)
+            
+            if mask.crs is None:
+                raise ValueError(
+                    "The mask must have a Coordinate Reference System (CRS)"
+                )
+            mask = mask.to_crs(self.zone_crs[zone_index]).geometry[0]
+            zone_geometry = zone_geometry.intersection(mask)
 
             # Consider only 8 decimal of precision
             # This is helpful to avoid floating point errors
@@ -210,27 +217,27 @@ class Equi7Grid:
             )
 
         # Obtain the bounding box of the specified zone
-        bbox_geo = zone_geometry.bounds
-        grid_user_distance = self.min_grid_size * (2 ** level)
+        bbox_geo: Tuple[float, float, float, float] = zone_geometry.bounds
+        grid_user_distance: int = self.min_grid_size * (2**level)
 
         # coordinate bottom left
         x_min, y_min, x_max, y_max = bbox_geo
 
         # Get the rounded coordinates (Xmin)
-        diff_x_min = x_min % grid_user_distance
-        x_init_rounded = x_min - diff_x_min
+        diff_x_min: float = x_min % grid_user_distance
+        x_init_rounded: float = x_min - diff_x_min
 
         # Get the rounded coordinates (Ymin)
-        diff_y_min = y_min % grid_user_distance
-        y_init_rounded = y_min - diff_y_min
+        diff_y_min: float = y_min % grid_user_distance
+        y_init_rounded: float = y_min - diff_y_min
 
         # Get the rounded coordinates (Xmax)
-        diff_x_max = x_max % grid_user_distance
-        x_last_rounded = x_max - diff_x_max + grid_user_distance
+        diff_x_max: float = x_max % grid_user_distance
+        x_last_rounded: float = x_max - diff_x_max + grid_user_distance
 
         # Get the rounded coordinates (Ymax)
-        diff_y_max = y_max % grid_user_distance
-        y_last_rounded = y_max - diff_y_max + grid_user_distance
+        diff_y_max: float = y_max % grid_user_distance
+        y_last_rounded: float = y_max - diff_y_max + grid_user_distance
 
         # Create the grid
         nxtiles = int((x_last_rounded - x_init_rounded) / grid_user_distance)
@@ -249,101 +256,112 @@ class Equi7Grid:
 
         # Create the final grid
         local_grid = gpd.GeoDataFrame(geometry=new_polys, crs=self.zone_crs[zone_index])
-        local_grid.insert(
-            0, "id", utils_add_id(local_grid, distance=self.min_grid_size, zone_id=zone)
-        )
-        local_grid.insert(1, "zone", zone)
-        local_grid.insert(2, "level", f"Z{level}")
+
+        # Add the zone and level
+        local_grid.insert(loc=0, column="zone", value=zone)
+        local_grid.insert(loc=1, column="level",value=f"Z{level}")
+
+        # Add the land column
+        land = self.zone_land[zone_index]
+        local_grid.insert(loc=2, column="land", value=local_grid.intersects(land).values)
 
         # intersect the grid with the zone geometry
-        return local_grid[local_grid.intersects(zone_geometry)]
+        final_result = local_grid[local_grid.intersects(zone_geometry)]
+        final_result.reset_index(drop=True, inplace=True)
+
+        return final_result
 
     def lonlat2grid(
-        self,
-        lon: float,
-        lat: float,
-        level: Optional[int] = 0,
-        centroid: Optional[bool] = True
+        self, lon: float, lat: float, centroid: Optional[bool] = True
     ) -> gpd.GeoDataFrame:
-        """Convert a latitude and longitude to an Equi7Grid Tile.
+        """Convert a latitude and longitude to an Equi7Grid Tile Metadata.
 
         Args:
-            lon (float): The longitude.
-            lat (float): The latitude.
-            level (Optional[int], optional): The grid level. Defaults to 0.
-            centroid (Optional[bool], optional): If True, it will return the
-                centroid of the grid tile. Defaults to True.
+            lon (float): The longitude in EPSG:4326.
+            lat (float): The latitude in EPSG:4326.
+            centroid (Optional[bool], optional): If True, it will align
+                with respect to the centroid of the grid tile. Defaults to
+                True.
 
         Returns:
-            gpd.GeoDataFrame: The Equi7Grid Tile.
+            gpd.GeoDataFrame: The Equi7Grid tile metadata for the
+                specified latitude and longitude.
         """
-        
-        # Create a gp dataframe
-        point = gpd.GeoDataFrame(geometry=[Point(lon, lat)], crs="EPSG:4326")
+
+        # Create a GeoDataFrame with the point
+        point = gpd.GeoDataFrame(
+            geometry=[shapely.geometry.Point(lon, lat)], crs="EPSG:4326"
+        )
 
         # Find the zone where this point is located
-        haversine_distance_min = math.inf
-        for index, zone in enumerate(self.zone_names):            
-            # Load the Zone
-            zone_geom = self.zone_geometry[index]
-            zone_crs = self.zone_crs[index]
-            zone_gpd = gpd.GeoDataFrame(geometry=[zone_geom], crs=self.zone_crs[index])
-            zone_geom = zone_gpd.geometry[0]
+        haversine_distance_min: float = math.inf
+        for index, zone in enumerate(self.zone_names):
 
+            # Load the Zone
+            zone_geom: shapely.geometry.Polygon = self.zone_geometry[index]
+            zone_crs: str = self.zone_crs[index] # in PROJCS format
+            zone_geom = gpd.GeoDataFrame(geometry=[zone_geom], crs=self.zone_crs[index]).geometry
+            
             # Intersect the point with the zone
-            condition = point.to_crs(zone_crs).intersects(zone_geom)[0]
+            condition: bool = point.to_crs(zone_crs).intersects(zone_geom)[0]
 
             # If the point is within many zones, select the closest one
             # to the origin of the zone
             if condition:
                 lon_ref, lat_ref = self.zone_origin[index]
-                arc_distance = haversine_distance(lon, lat, lon_ref, lat_ref)
-                
+
+                # Calculate the arc distance between the point and the origin of the Equi7Grid zone
+                arc_distance: float = haversine_distance(lon, lat, lon_ref, lat_ref)
+
                 # Update the minimum distance and set the best index (best zone)
                 if arc_distance < haversine_distance_min:
                     haversine_distance_min = arc_distance
                     best_index = index
 
         # Get the Equi7Grid Zone name
-        name = self.zone_names[best_index]
+        name: str = self.zone_names[best_index]
 
-        # Search in the level 1 grid & add bottom left coordinates
-        q = self.create_grid(level=level, zone=name, mask=point.geometry[0], coverland=False)
-        
+        # Search in the level 0 grid & add reference coordinates
+        q: gpd.GeoDataFrame = self.create_grid(level=0, zone=name, mask=point.geometry)
+
+        # add the Equi7Grid tile ID
+        q_id: str = get_id(polygon=q.geometry.values[0], distance=self.min_grid_size, zone_id=name)
+        q.insert(loc=0, column="id", value=q_id)
+
         if centroid:
-            q.insert(1, "x", q.geometry.centroid.x)
-            q.insert(2, "y", q.geometry.centroid.y)
+            x, y = q.geometry.centroid.x.values[0], q.geometry.centroid.y.values[0]
+            q.insert(loc=1, column="x", value=x)
+            q.insert(loc=2, column="y", value=y)
         else:
-            q.insert(1, "x", q.geometry.bounds.minx)
-            q.insert(2, "y", q.geometry.bounds.miny)        
+            x, y = q.geometry.bounds.minx.values[0], q.geometry.bounds.miny.values[0]
+            q.insert(loc=1, column="x", value=x)
+            q.insert(loc=2, column="y", value=y)
+
+        # Add the lonlat coordinates
+        lon, lat = pyproj.Transformer.from_crs(
+            crs_from=q.crs, crs_to="EPSG:4326", always_xy=True
+        ).transform(x, y)
+        q.insert(loc=1, column="lon", value=lon)
+        q.insert(loc=2, column="lat", value=lat)
 
         return q
 
     def grid2lonlat(
-        self,
-        grid_id: Union[str, gpd.GeoDataFrame],
-        xy_coords: Optional[bool] = False,
-        centroid: Optional[bool] = True
+        self, grid_id: str, centroid: Optional[bool] = True
     ) -> pd.DataFrame:
-        """Convert an Equi7Grid Tile to a latitude and longitude.
+        """Convert an Equi7Grid grid id to a Equi7Grid Tile Metadata. 
 
         Args:
-            grid_id (str): The Equi7Grid Tile ID.
-            xy_coords (Optional[bool], optional): If True, it will
-                also return the X and Y coordinates in the Equi7Grid
-                system. Defaults to True.
-            centroid (Optional[bool], optional): If True, it will 
-                return the centroid of the grid tile. Defaults to 
+            grid_id (str): The Equi7Grid grid id.
+            centroid (Optional[bool], optional): If True, it will align
+                with respect to the centroid of the grid tile. Defaults to
                 True.
 
         Returns:
-            pd.DataFrame: A DataFrame with the latitude and longitude
+            pd.DataFrame: A DataFrame with the Equi7Grid tile metadata
         """
-        # Check if the grid_id is a string or a GeoDataFrame
-        if isinstance(grid_id, gpd.GeoDataFrame):
-            grid_id = grid_id.id.values[0]
-            
-        ## Extract the metadata from the grid_id
+
+        # Extract the metadata from the grid_id
         re_expr = re.compile(r"\b([A-Z]+)(\d+)_E(\d+)N(\d+)")
         zone = re_expr.search(grid_id).group(1)
         distance = int(re_expr.search(grid_id).group(2))
@@ -352,98 +370,188 @@ class Equi7Grid:
 
         # From Grid to Equi7Grid coordinates
         if centroid:
-            x = (nxtile + 0.5) * distance
-            y = (nytile + 0.5) * distance
+            x: float = (nxtile + 0.5) * distance
+            y: float = (nytile + 0.5) * distance
         else:
-            x = nxtile * distance
-            y = nytile * distance
+            x: float = nxtile * distance
+            y: float = nytile * distance
 
-        point_local = Point(x, y)
+        # Get the grid CRS
+        crs: str = self.zone_crs[self.zone_names.index(zone)]
+        lon, lat = pyproj.Transformer.from_crs(
+            crs_from=crs, crs_to="EPSG:4326", always_xy=True
+        ).transform(x, y)
 
-        # From Equi7Grid to Geographic coordinates
-        zone_crs = self.zone_crs[self.zone_names.index(zone)]
-        point_geo = gpd.GeoDataFrame(geometry=[point_local], crs=zone_crs)
-        point_geo = point_geo.to_crs("EPSG:4326").geometry[0]
+        return self.lonlat2grid(lon=lon, lat=lat, centroid=centroid)
 
-        # Return the coordinates
-        if xy_coords:
-            results = {"lon": point_geo.x, "lat": point_geo.y, "x": x, "y": y}
-        else:
-            results = {"lon": point_geo.x, "lat": point_geo.y}
+    def zone_metadata(self, zone: Literal["AN", "NA", "OC", "SA", "AF", "EU", "AS"]) -> Equi7GridZone:
+        """Obtain the metadata of a specific Equi7Grid zone.
 
-        # from dict to dataframe
-        return pd.DataFrame(results, index=[0])
+        Args:
+            zone (str): The Equi7Grid zone.
 
-    def align2grid(
-        self,
-        lon: float,
-        lat: float,    
-        level: Optional[int] = 0,
-        centroid: Optional[bool] = True
+        Returns:
+            Equi7GridZone: The metadata of the Equi7Grid zone.
+        """
+        return getattr(self, zone)
+
+    def grid2grid(
+        self, poly: shapely.geometry.Polygon, region_code: str
     ) -> gpd.GeoDataFrame:
-        """ Align the grid to the nearest grid tile.
+        """Given a superior grid tile, return the subgrids at
+        the finest level. This function runs in parallel to
+        speed up the process and do not run validations.
 
         Args:
-            lon (float): The longitude.
-            lat (float): The latitude.
-            level (int): The grid level. Defaults to 0.
-            centroid (Optional[bool], optional): If True, it will 
-                return the centroid of the grid tile. Defaults to
+            poly (shapely.geometry.Polygon): The superior grid tile.
+            region_code (str): The region code of the grid tile.
+            centroid (Optional[bool], optional): If True, it will align
+                with respect to the centroid of the grid tile. Defaults to
                 True.
-
         Returns:
-            gpd.GeoDataFrame: A GeoDataFrame with the coordinates
-                aligned to the grid.
+            gpd.GeoDataFrame: The subgrids at the finest level.
         """
-        return self.grid2lonlat(
-            grid_id=self.lonlat2grid(lon, lat, level=level, centroid=centroid),
-            centroid=centroid,
-            xy_coords=False
-        )
-
-    def lonlat2grid_ids(
-        self,
-        lon: float,
-        lat: float,
-        level: int
-    ) -> list:
-        """Obtain the grid ids from a given point.
-
-        Args:
-            lon (float): Longitude of the point.
-            lat (float): Latitude of the point.
-            level (int): Level of the grid.
-
-        Returns:
-            list: A list with the grid ids.
-        """
-        # Obtain the boundaries of the grid
-        grid_information = self.lonlat2grid(lon=lon, lat=lat, level=level)
 
         # Obtain the boundaries of the grid
-        minx, miny, maxx, maxy = grid_information.geometry.bounds.values[0]
-        
-        # Obtain the region code and the distance
-        re_expr = re.compile(r"\b([A-Z]+)(\d+)_E(\d+)N(\d+)")
-        region_code = re_expr.match(grid_information.id.values[0]).group(1)
-        distance = int(re_expr.match(grid_information.id.values[0]).group(2))        
+        minx, miny, maxx, maxy = poly.bounds
+        distance: int = self.min_grid_size
 
         # Calculate the number of steps in the x and y directions
-        x_steps = math.ceil((maxx - minx) / distance)
-        y_steps = math.ceil((maxy - miny) / distance)
-        
-        grids = []    
+        x_steps: int = math.ceil((maxx - minx) / distance)
+        y_steps: int = math.ceil((maxy - miny) / distance)
+
+        # Project the Equi7Grid to Geographic coordinates
+        poly_crs: str = self.zone_metadata(region_code).crs # in PROJCS format
+        reprojector = pyproj.Transformer.from_crs(
+            crs_from=poly_crs, crs_to="EPSG:4326", always_xy=True
+        )
+
+        # Initialize an empty list to store the DataFrames for each grid
+        grids = []
         for i in range(x_steps):
+            # Create a new DataFrame for the current x step
+            grids_j = pd.DataFrame()
+
+            # Initialize lists to store x, y coordinates and names for the current grid
+            x_list = []
+            y_list = []
+            name_list = []
+            bbox_list = []
+
+            # Iterate over the number of y steps
             for j in range(y_steps):
+                # Calculate the x and y coordinates for the current grid cell
+                x = minx + i * distance + distance // 2
+                y = miny + j * distance + distance // 2
 
                 # Calculate the boundaries of each grid cell
-                cell_minx = str(int(minx / distance + i * distance))
-                cell_miny = str(int(miny / distance + j * distance))
-                
-                # Cook the grid name
-                name = f"{region_code}{distance}_E{cell_minx}N{cell_miny}"                
-                grids.append(name)
-        return grids
+                cell_minx = str(int(x / distance)).zfill(4)
+                cell_miny = str(int(y / distance)).zfill(4)
+
+                # calculate bbox
+                bbox: List[float] = [
+                    x - distance // 2,
+                    y - distance // 2,
+                    x + distance // 2,
+                    y + distance // 2,
+                ]
+                tile = shapely.geometry.box(*bbox)
+
+                # Generate the unique id for the current grid cell
+                name = f"{region_code}{distance}_E{cell_minx}N{cell_miny}"
+                name_list.append(name)
+                x_list.append(x)
+                y_list.append(y)
+                bbox_list.append(tile)
+
+            # Assign the lists to the DataFrame columns
+            grids_j["id"] = name_list
+            grids_j["x"] = x_list
+            grids_j["y"] = y_list
+
+            # Convert Equi7Grid coordinates to geographic coordinates (longitude, latitude)
+            grids_j["lon"], grids_j["lat"] = reprojector.transform(
+                grids_j["x"], grids_j["y"]
+            )
+
+            # Add the region code as the zone for the current grid
+            grids_j["zone"] = region_code
+
+            # Add the geometry column
+            grids_j["geometry"] = bbox_list
+
+            # Append the DataFrame for the current grid to the list of grids
+            grids.append(grids_j)
+
+        # Concatenate the DataFrames for each grid cell
+        dataset = gpd.GeoDataFrame(pd.concat(grids, ignore_index=True), crs=poly_crs)
+
+        # Intersect the grid with the land
+        land: shapely.geometry.MultiPolygon = self.zone_metadata(region_code).landmasses_equi7grid
+
+        # Use multiprocessing to speed up the process
+        with mp.Pool(mp.cpu_count()) as pool:
+            intersects_func_partial = partial(intersects_func, land=land)
+            dataset["land"] = pool.starmap(
+                intersects_func_partial,
+                zip(
+                    dataset["geometry"],
+                ),
+            )
+
+        return dataset
+
+    def cubo_utm_parameters(
+        self, lon: float, lat: float, rotation_compensation: float = 1.1
+    ) -> Dict[str, float]:
+        """Get the parameters for a CUBO request given a lonlat
+
+        Args:
+            lon (float): The longitude of the point in EPSG:4326.
+            lat (float): The latitude of the point in EPSG:4326.
+            rotation_compensation (float, optional): Compensation 
+                factor to avoid border effects. Defaults to 1.1.
+
+        Returns:
+            dict: A dictionary with the parameters for a 
+                CUBO request. The dictionary contains the
+                following keys:
+                - lon: The fixed longitude of the point.
+                - lat: The fixed latitude of the point.
+                - utm: The UTM EPSG code.
+                - distance: The distance of the CUBO request.
+                - polygon: The POLYGON in the UTM coordinates.
+                - zone: The Equi7Grid zone.
+                - crs: The CRS of the zone.                
+        """
+
+        # Obtain the new coordinates and POLYGON
+        gmeta: gpd.GeoDataFrame = self.lonlat2grid(lon, lat)
+
+        # Get the UTM EPSG from latlon
+        epsg: str = pyproj.database.query_utm_crs_info(
+            datum_name="WGS 84",
+            area_of_interest=pyproj.aoi.AreaOfInterest(lon, lat, lon, lat),
+        )[0].code
+        utm_crs: pyproj.CRS = pyproj.CRS.from_epsg(epsg)
+
+        # Convert the polygon to UTM and obtain max distance
+        gpoly: Tuple[float, float, float, float] = gmeta["geometry"].to_crs(epsg).values[0].bounds
+        x_dist: float = gpoly[2] - gpoly[0]
+        y_dist: float = gpoly[3] - gpoly[1]
+
+        # add a extra distance to avoid border effects
+        distance = int(np.ceil(max(x_dist, y_dist) * rotation_compensation))
+
+        return {
+            "lon": gmeta["lon"].values[0],
+            "lat": gmeta["lat"].values[0],
+            "utm": "EPSG:{}".format(utm_crs.to_epsg()),
+            "distance": distance,
+            "polygon": gmeta["geometry"].values[0],
+            "zone": gmeta["zone"].values[0],
+            "crs": self.__dict__.get(gmeta["zone"].values[0]).crs,
+        }
 
     def __str__(self) -> str:
         """Display the Equi7Grid information.